refactor(server): split route islands
