+++ noor-server/src/server/routes.rs
@@ -32,11 +32,14 @@ use rusqlite::{OptionalExtension, params};
 use serde::{Deserialize, Serialize};
 use serde_json::{Value, json};
 use std::collections::{HashMap, HashSet};
-use std::sync::{Arc, Mutex as StdMutex, OnceLock};
+use std::sync::Arc;
 use std::time::{Duration, Instant};
 use tracing::{error, info, warn};
 
+mod analytics_routes;
+mod chart_routes;
 mod genre_routes;
+mod sportify_routes;
 
 #[derive(Debug, Deserialize)]
 pub struct ListParams {
@@ -64,23 +67,6 @@ pub struct ListParams {
 pub struct SearchParams {
     q: String,
     limit: Option<i64>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct ListenHistoryParams {
-    limit: Option<i64>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct AnalyticsDashboardParams {
-    recent_limit: Option<i64>,
-    top_limit: Option<i64>,
-    days: Option<i64>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct AnalyticsSignalsParams {
-    days: Option<i64>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -359,10 +345,22 @@ pub fn api_routes(state: SharedState) -> Router {
             "/api/smart/playlists/{id}/evaluate",
             get(evaluate_smart_playlist),
         )
-        .route("/api/analytics/overview", get(get_analytics_overview))
-        .route("/api/analytics/dashboard", get(get_analytics_dashboard))
-        .route("/api/analytics/signals", get(get_analytics_signals))
-        .route("/api/analytics/listens/recent", get(get_recent_listens))
+        .route(
+            "/api/analytics/overview",
+            get(analytics_routes::get_analytics_overview),
+        )
+        .route(
+            "/api/analytics/dashboard",
+            get(analytics_routes::get_analytics_dashboard),
+        )
+        .route(
+            "/api/analytics/signals",
+            get(analytics_routes::get_analytics_signals),
+        )
+        .route(
+            "/api/analytics/listens/recent",
+            get(analytics_routes::get_recent_listens),
+        )
         .route("/api/discovery/preview", post(preview_discovery))
         .route("/api/discovery/new", post(discover_new_music))
         .route("/api/discovery/save", post(save_discovery_track))
@@ -413,44 +411,47 @@ pub fn api_routes(state: SharedState) -> Router {
         // affected.
         .route(
             "/api/discovery/sportify/search",
-            get(sportify_discovery_search),
+            get(sportify_routes::sportify_discovery_search),
         )
         .route(
             "/api/discovery/sportify/track/{spotify_id}",
-            get(sportify_discovery_track),
+            get(sportify_routes::sportify_discovery_track),
         )
         .route(
             "/api/discovery/sportify/album/{spotify_id}",
-            get(sportify_discovery_album),
+            get(sportify_routes::sportify_discovery_album),
         )
         .route(
             "/api/discovery/sportify/playlist/{spotify_id}",
-            get(sportify_discovery_playlist),
+            get(sportify_routes::sportify_discovery_playlist),
         )
         .route(
             "/api/discovery/sportify/artist/{spotify_id}",
-            get(sportify_discovery_artist),
+            get(sportify_routes::sportify_discovery_artist),
         )
         .route(
             "/api/discovery/sportify/artist/{spotify_id}/top-tracks",
-            get(sportify_discovery_artist_top_tracks),
+            get(sportify_routes::sportify_discovery_artist_top_tracks),
         )
         .route(
             "/api/discovery/sportify/artist/{spotify_id}/related",
-            get(sportify_discovery_artist_related),
+            get(sportify_routes::sportify_discovery_artist_related),
         )
         .route(
             "/api/discovery/sportify/album/{spotify_id}/related",
-            get(sportify_discovery_album_related),
+            get(sportify_routes::sportify_discovery_album_related),
         )
         .route(
             "/api/discovery/sportify/track/{spotify_id}/related",
-            get(sportify_discovery_track_related),
+            get(sportify_routes::sportify_discovery_track_related),
         )
         // Save an ephemeral Spotify-sourced playlist into the user's library.
         // Imports each resolved TIDAL track + creates a noor playlist; rows
         // without a TIDAL match are skipped (counted in the response).
-        .route("/api/spotify-playlist/save", post(save_spotify_playlist))
+        .route(
+            "/api/spotify-playlist/save",
+            post(sportify_routes::save_spotify_playlist),
+        )
         .route("/api/radio/song", post(radio_song))
         .route("/api/radio/album", post(radio_album))
         .route("/api/radio/artist", post(radio_artist))
@@ -668,9 +669,15 @@ pub fn api_routes(state: SharedState) -> Router {
             get(get_tidal_discover_module_items),
         )
         // Trending / charts (Phase 5)
-        .route("/api/charts", get(get_charts))
-        .route("/api/charts/lastfm/genres", get(list_lastfm_genres))
-        .route("/api/charts/lastfm/countries", get(list_lastfm_countries))
+        .route("/api/charts", get(chart_routes::get_charts))
+        .route(
+            "/api/charts/lastfm/genres",
+            get(chart_routes::list_lastfm_genres),
+        )
+        .route(
+            "/api/charts/lastfm/countries",
+            get(chart_routes::list_lastfm_countries),
+        )
         // Server auth management
         .route("/api/server/token", get(get_server_token_handler))
         .route(
@@ -1808,82 +1815,6 @@ async fn delete_smart_playlist_route(
             };
             (status, Json(json!({ "message": e.to_string() })))
         })
-}
-
-async fn get_analytics_overview(
-    State(state): State<SharedState>,
-) -> Result<Json<Value>, StatusCode> {
-    let state = state.read().await;
-    state
-        .db
-        .with_conn(|conn| {
-            let overview = queries::get_analytics_overview(conn)?;
-            Ok(Json(json!({ "overview": overview })))
-        })
-        .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)
-}
-
-async fn get_recent_listens(
-    State(state): State<SharedState>,
-    Query(params): Query<ListenHistoryParams>,
-) -> Result<Json<Value>, StatusCode> {
-    let limit = params.limit.unwrap_or(25).clamp(1, 200);
-    let state = state.read().await;
-    state
-        .db
-        .with_conn(|conn| {
-            let listens = queries::get_recent_listens(conn, limit)?;
-            Ok(Json(json!({ "listens": listens })))
-        })
-        .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)
-}
-
-async fn get_analytics_dashboard(
-    State(state): State<SharedState>,
-    Query(params): Query<AnalyticsDashboardParams>,
-) -> Result<Json<Value>, StatusCode> {
-    let recent_limit = params.recent_limit.unwrap_or(12).clamp(1, 50);
-    let top_limit = params.top_limit.unwrap_or(8).clamp(1, 20);
-    // Relaxed from 90 → 36500 so the new analytics page's "All" pill can pass through to
-    // the legacy dashboard endpoint without silent truncation. Existing callers passing
-    // ≤90 are unaffected.
-    let days = params.days.unwrap_or(14).clamp(1, 36500);
-
-    let state = state.read().await;
-    let dashboard = state
-        .db
-        .with_conn(|conn| {
-            Ok(crate::db::models::AnalyticsDashboard {
-                overview: queries::get_analytics_overview(conn)?,
-                recent_listens: queries::get_recent_listens(conn, recent_limit)?,
-                top_tracks: queries::get_top_tracks_by_history(conn, top_limit)?,
-                top_artists: queries::get_top_artists_by_history(conn, top_limit)?,
-                top_genres: queries::get_top_genres_by_history(conn, top_limit)?,
-                activity: queries::get_listen_activity(conn, days)?,
-                behavior: queries::get_behavior_metrics(conn)?,
-            })
-        })
-        .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?;
-
-    Ok(Json(json!({ "dashboard": dashboard })))
-}
-
-async fn get_analytics_signals(
-    State(state): State<SharedState>,
-    Query(params): Query<AnalyticsSignalsParams>,
-) -> Result<Json<Value>, StatusCode> {
-    let days = params.days.unwrap_or(30).clamp(1, 36500);
-
-    let state = state.read().await;
-    let signals = state
-        .db
-        .with_conn(|conn| queries::get_analytics_signals(conn, days))
-        .map_err(|err| {
-            tracing::error!(?err, days, "get_analytics_signals failed");
-            StatusCode::INTERNAL_SERVER_ERROR
-        })?;
-
-    Ok(Json(json!({ "signals": signals })))
 }
 
 async fn preview_discovery(
@@ -3703,789 +3634,6 @@ fn persist_outcome(
             let _ = sp_cache::put_unresolved(conn, spotify_id, Some(&outcome.reason));
         }
     }
-}
-
-#[derive(Debug, Deserialize)]
-struct SportifySearchQuery {
-    q: String,
-    #[serde(default)]
-    r#type: Option<String>,
-    #[serde(default)]
-    limit: Option<u32>,
-    #[serde(default)]
-    offset: Option<u32>,
-}
-
-fn parse_search_kind(s: Option<&str>) -> crate::services::sportify::client::SportifySearchKind {
-    use crate::services::sportify::client::SportifySearchKind;
-    match s.map(str::to_ascii_lowercase).as_deref() {
-        Some("album") => SportifySearchKind::Album,
-        Some("artist") => SportifySearchKind::Artist,
-        Some("playlist") => SportifySearchKind::Playlist,
-        // Default to track when missing or unknown — matches Sportify's own
-        // most-common usage.
-        _ => SportifySearchKind::Track,
-    }
-}
-
-async fn sportify_discovery_search(
-    State(state): State<SharedState>,
-    Query(params): Query<SportifySearchQuery>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{cache as sp_cache, normalize};
-
-    let q = params.q.trim();
-    if q.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "q required" })),
-        ));
-    }
-
-    let kind = parse_search_kind(params.r#type.as_deref());
-    let limit = params.limit.unwrap_or(20).clamp(1, 50);
-    let offset = params.offset.unwrap_or(0);
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    let cached = db
-        .with_conn(|conn| sp_cache::get_search(conn, &cache_cfg, q, kind, limit, offset))
-        .map_err(internal)?;
-
-    let payload = match cached {
-        Some(p) => p,
-        None => {
-            let fetched = sportify_client
-                .search(q, kind, limit, offset)
-                .await
-                .map_err(|e| {
-                    (
-                        StatusCode::BAD_GATEWAY,
-                        Json(json!({ "error": format!("sportify_search: {e}") })),
-                    )
-                })?;
-            db.with_conn(|conn| sp_cache::put_search(conn, q, kind, limit, offset, &fetched))
-                .map_err(internal)?;
-            fetched
-        }
-    };
-
-    let mut normalized = normalize::search_from_sportify(&payload, "sportify_search");
-    db.with_conn(|conn| {
-        normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut normalized.tracks)?;
-        for album in normalized.albums.iter_mut() {
-            normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut album.tracks)?;
-        }
-        for playlist in normalized.playlists.iter_mut() {
-            normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut playlist.tracks)?;
-        }
-        Ok(())
-    })
-    .map_err(internal)?;
-
-    Ok(Json(serde_json::to_value(normalized).unwrap_or(json!({}))))
-}
-
-async fn sportify_discovery_track(
-    State(state): State<SharedState>,
-    Path(spotify_id): Path<String>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{cache as sp_cache, normalize};
-
-    let id = spotify_id.trim();
-    if id.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "spotify_id required" })),
-        ));
-    }
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    let track = match db
-        .with_conn(|conn| sp_cache::get_track_meta(conn, &cache_cfg, id))
-        .map_err(internal)?
-    {
-        Some(t) => t,
-        None => {
-            let fetched = sportify_client.track(id).await.map_err(|e| {
-                (
-                    StatusCode::BAD_GATEWAY,
-                    Json(json!({ "error": format!("sportify_track_fetch: {e}") })),
-                )
-            })?;
-            db.with_conn(|conn| {
-                sp_cache::put_track_meta(conn, id, &fetched)?;
-                crate::services::sportify::stats::write_track_playcount(conn, &fetched);
-                Ok::<_, anyhow::Error>(())
-            })
-            .map_err(internal)?;
-            fetched
-        }
-    };
-
-    let mut row = normalize::track_from_sportify(&track, "sportify_track");
-    db.with_conn(|conn| {
-        normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, std::slice::from_mut(&mut row))
-    })
-    .map_err(internal)?;
-
-    Ok(Json(serde_json::to_value(row).unwrap_or(json!({}))))
-}
-
-async fn sportify_discovery_album(
-    State(state): State<SharedState>,
-    Path(spotify_id): Path<String>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{cache as sp_cache, normalize};
-
-    let id = spotify_id.trim();
-    if id.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "spotify_id required" })),
-        ));
-    }
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    let album = match db
-        .with_conn(|conn| sp_cache::get_album_meta(conn, &cache_cfg, id))
-        .map_err(internal)?
-    {
-        Some(a) => a,
-        None => {
-            let fetched = sportify_client.album(id).await.map_err(|e| {
-                (
-                    StatusCode::BAD_GATEWAY,
-                    Json(json!({ "error": format!("sportify_album_fetch: {e}") })),
-                )
-            })?;
-            db.with_conn(|conn| {
-                sp_cache::put_album_meta(conn, id, &fetched)?;
-                crate::services::sportify::stats::write_track_playcounts(conn, &fetched.tracks);
-                Ok::<_, anyhow::Error>(())
-            })
-            .map_err(internal)?;
-            fetched
-        }
-    };
-
-    let mut row = normalize::album_from_sportify(&album, "sportify_album");
-    let pending_ids = eager_and_lazy_resolve_for_list(&state, &album.tracks).await;
-    db.with_conn(|conn| {
-        normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut row.tracks)
-    })
-    .map_err(internal)?;
-
-    Ok(Json(json!({
-        "album": serde_json::to_value(row).unwrap_or(json!({})),
-        "pendingSpotifyIds": pending_ids,
-    })))
-}
-
-async fn sportify_discovery_playlist(
-    State(state): State<SharedState>,
-    Path(spotify_id): Path<String>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{cache as sp_cache, normalize};
-
-    let id = spotify_id.trim();
-    if id.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "spotify_id required" })),
-        ));
-    }
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    let playlist = match db
-        .with_conn(|conn| sp_cache::get_playlist_meta(conn, &cache_cfg, id))
-        .map_err(internal)?
-    {
-        Some(p) => p,
-        None => {
-            let fetched = sportify_client.playlist(id).await.map_err(|e| {
-                (
-                    StatusCode::BAD_GATEWAY,
-                    Json(json!({ "error": format!("sportify_playlist_fetch: {e}") })),
-                )
-            })?;
-            db.with_conn(|conn| {
-                sp_cache::put_playlist_meta(conn, id, &fetched)?;
-                crate::services::sportify::stats::write_track_playcounts(conn, &fetched.tracks);
-                Ok::<_, anyhow::Error>(())
-            })
-            .map_err(internal)?;
-            fetched
-        }
-    };
-
-    let mut row = normalize::playlist_from_sportify(&playlist, "sportify_playlist");
-    db.with_conn(|conn| {
-        normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut row.tracks)
-    })
-    .map_err(internal)?;
-    let pending_ids = spawn_background_resolve_for_list(&state, &playlist.tracks).await;
-
-    Ok(Json(json!({
-        "playlist": serde_json::to_value(row).unwrap_or(json!({})),
-        "pendingSpotifyIds": pending_ids,
-    })))
-}
-
-async fn sportify_discovery_artist(
-    State(state): State<SharedState>,
-    Path(spotify_id): Path<String>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{cache as sp_cache, normalize};
-
-    let id = spotify_id.trim();
-    if id.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "spotify_id required" })),
-        ));
-    }
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    let artist = match db
-        .with_conn(|conn| sp_cache::get_artist_meta(conn, &cache_cfg, id))
-        .map_err(internal)?
-    {
-        Some(a) => a,
-        None => {
-            let fetched = sportify_client.artist(id).await.map_err(|e| {
-                (
-                    StatusCode::BAD_GATEWAY,
-                    Json(json!({ "error": format!("sportify_artist_fetch: {e}") })),
-                )
-            })?;
-            db.with_conn(|conn| {
-                sp_cache::put_artist_meta(conn, id, &fetched)?;
-                crate::services::sportify::stats::write_artist_monthly_listeners(conn, &fetched);
-                Ok::<_, anyhow::Error>(())
-            })
-            .map_err(internal)?;
-            fetched
-        }
-    };
-
-    let row = normalize::artist_from_sportify(&artist);
-    Ok(Json(serde_json::to_value(row).unwrap_or(json!({}))))
-}
-
-async fn sportify_discovery_artist_top_tracks(
-    State(state): State<SharedState>,
-    Path(spotify_id): Path<String>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{cache as sp_cache, normalize};
-
-    let id = spotify_id.trim();
-    if id.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "spotify_id required" })),
-        ));
-    }
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    // Top tracks aren't cached as a unit (they're a derived list); rely on
-    // the per-track meta cache to absorb repeat hits.
-    let tracks = sportify_client.artist_top_tracks(id).await.map_err(|e| {
-        (
-            StatusCode::BAD_GATEWAY,
-            Json(json!({ "error": format!("sportify_top_tracks_fetch: {e}") })),
-        )
-    })?;
-
-    db.with_conn(|conn| {
-        for t in &tracks {
-            if let Some(track_id) = t.id.as_deref() {
-                let _ = sp_cache::put_track_meta(conn, track_id, t);
-            }
-        }
-        crate::services::sportify::stats::write_track_playcounts(conn, &tracks);
-        Ok::<_, anyhow::Error>(())
-    })
-    .map_err(internal)?;
-
-    let mut rows: Vec<_> = tracks
-        .iter()
-        .map(|t| normalize::track_from_sportify(t, "sportify_artist_top_tracks"))
-        .collect();
-    let pending_ids = eager_and_lazy_resolve_for_list(&state, &tracks).await;
-    db.with_conn(|conn| normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut rows))
-        .map_err(internal)?;
-
-    Ok(Json(json!({
-        "spotifyId": id,
-        "tracks": rows,
-        "pendingSpotifyIds": pending_ids,
-    })))
-}
-
-async fn sportify_discovery_artist_related(
-    State(state): State<SharedState>,
-    Path(spotify_id): Path<String>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{normalize, recommend};
-
-    let id = spotify_id.trim();
-    if id.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "spotify_id required" })),
-        ));
-    }
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    let related = recommend::artist_related(&sportify_client, &db, &cache_cfg, id)
-        .await
-        .map_err(|e| {
-            (
-                StatusCode::BAD_GATEWAY,
-                Json(json!({ "error": format!("sportify_related: {e}") })),
-            )
-        })?;
-
-    // Track lists go through eager+lazy resolution; albums/artists carry no
-    // tidal state since they're just navigation cards.
-    let mut top_rows: Vec<_> = related
-        .top_tracks
-        .iter()
-        .map(|t| normalize::track_from_sportify(t, "sportify_artist_related_top"))
-        .collect();
-    let mut deep_rows: Vec<_> = related
-        .deep_cuts
-        .iter()
-        .map(|t| normalize::track_from_sportify(t, "sportify_artist_related_deep"))
-        .collect();
-
-    let mut pending = Vec::new();
-    pending.extend(eager_and_lazy_resolve_for_list(&state, &related.top_tracks).await);
-    pending.extend(eager_and_lazy_resolve_for_list(&state, &related.deep_cuts).await);
-
-    db.with_conn(|conn| {
-        normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut top_rows)?;
-        normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut deep_rows)?;
-        Ok::<_, anyhow::Error>(())
-    })
-    .map_err(internal)?;
-
-    let recent_releases: Vec<_> = related
-        .recent_releases
-        .iter()
-        .map(|a| normalize::album_from_sportify(a, "sportify_artist_related_recent"))
-        .collect();
-    let similar_artists: Vec<_> = related
-        .similar_artists
-        .iter()
-        .map(normalize::artist_from_sportify)
-        .collect();
-
-    Ok(Json(json!({
-        "spotifyId": id,
-        "topTracks": top_rows,
-        "deepCuts": deep_rows,
-        "recentReleases": recent_releases,
-        "similarArtists": similar_artists,
-        "pendingSpotifyIds": pending,
-    })))
-}
-
-async fn sportify_discovery_album_related(
-    State(state): State<SharedState>,
-    Path(spotify_id): Path<String>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{normalize, recommend};
-
-    let id = spotify_id.trim();
-    if id.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "spotify_id required" })),
-        ));
-    }
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    let related = recommend::album_related(&sportify_client, &db, &cache_cfg, id)
-        .await
-        .map_err(|e| {
-            (
-                StatusCode::BAD_GATEWAY,
-                Json(json!({ "error": format!("sportify_related: {e}") })),
-            )
-        })?;
-
-    let mut more_from_artist: Vec<_> = related
-        .more_from_artist
-        .iter()
-        .map(|t| normalize::track_from_sportify(t, "sportify_album_related"))
-        .collect();
-    let pending = eager_and_lazy_resolve_for_list(&state, &related.more_from_artist).await;
-
-    db.with_conn(|conn| {
-        normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut more_from_artist)
-    })
-    .map_err(internal)?;
-
-    let more_albums_by_artist: Vec<_> = related
-        .more_albums_by_artist
-        .iter()
-        .map(|a| normalize::album_from_sportify(a, "sportify_album_related_albums"))
-        .collect();
-
-    Ok(Json(json!({
-        "spotifyId": id,
-        "moreFromArtist": more_from_artist,
-        "moreAlbumsByArtist": more_albums_by_artist,
-        "pendingSpotifyIds": pending,
-    })))
-}
-
-async fn sportify_discovery_track_related(
-    State(state): State<SharedState>,
-    Path(spotify_id): Path<String>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{normalize, recommend};
-
-    let id = spotify_id.trim();
-    if id.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "spotify_id required" })),
-        ));
-    }
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    let related = recommend::track_related(&sportify_client, &db, &cache_cfg, id)
-        .await
-        .map_err(|e| {
-            (
-                StatusCode::BAD_GATEWAY,
-                Json(json!({ "error": format!("sportify_related: {e}") })),
-            )
-        })?;
-
-    let mut more_from_album: Vec<_> = related
-        .more_from_album
-        .iter()
-        .map(|t| normalize::track_from_sportify(t, "sportify_track_related_album"))
-        .collect();
-    let mut more_from_artist: Vec<_> = related
-        .more_from_artist
-        .iter()
-        .map(|t| normalize::track_from_sportify(t, "sportify_track_related_artist"))
-        .collect();
-
-    let mut pending = Vec::new();
-    pending.extend(eager_and_lazy_resolve_for_list(&state, &related.more_from_album).await);
-    pending.extend(eager_and_lazy_resolve_for_list(&state, &related.more_from_artist).await);
-
-    db.with_conn(|conn| {
-        normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut more_from_album)?;
-        normalize::enrich_tracks_with_tidal_cache(conn, &cache_cfg, &mut more_from_artist)?;
-        Ok::<_, anyhow::Error>(())
-    })
-    .map_err(internal)?;
-
-    Ok(Json(json!({
-        "spotifyId": id,
-        "moreFromAlbum": more_from_album,
-        "moreFromArtist": more_from_artist,
-        "pendingSpotifyIds": pending,
-    })))
-}
-
-#[derive(Debug, Deserialize)]
-struct SaveSpotifyPlaylistBody {
-    spotify_id: String,
-    /// Override for the noor playlist name. Defaults to the Sportify
-    /// playlist title.
-    #[serde(default)]
-    name: Option<String>,
-}
-
-/// Save an ephemeral Spotify-sourced playlist into the user's library.
-///
-/// Pre-condition: the playlist's tracks have been bulk-resolved against
-/// TIDAL (the ephemeral view does this on open). Tracks without a cached
-/// resolution are skipped — we never invent a placeholder TIDAL id.
-async fn save_spotify_playlist(
-    State(state): State<SharedState>,
-    Json(body): Json<SaveSpotifyPlaylistBody>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::sportify::{cache as sp_cache, recommend};
-
-    let id = body.spotify_id.trim();
-    if id.is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "spotify_id required" })),
-        ));
-    }
-
-    let (sportify_client, cache_cfg, db) = {
-        let s = state.read().await;
-        (
-            s.sportify_client.clone(),
-            s.sportify_cache_config,
-            s.db.clone(),
-        )
-    };
-    let Some(sportify_client) = sportify_client else {
-        return Err((
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "sportify_unavailable" })),
-        ));
-    };
-
-    let playlist = recommend::cached_playlist(&sportify_client, &db, &cache_cfg, id)
-        .await
-        .map_err(|e| {
-            (
-                StatusCode::BAD_GATEWAY,
-                Json(json!({ "error": format!("sportify_playlist_fetch: {e}") })),
-            )
-        })?;
-
-    let playlist_name = body
-        .name
-        .as_deref()
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .map(|s| s.to_string())
-        .or_else(|| playlist.name.clone())
-        .unwrap_or_else(|| "Spotify playlist".to_string());
-
-    // Pull cached TIDAL resolutions for the playlist's tracks. Tracks without
-    // a cached hit are skipped here — the frontend should bulk-resolve before
-    // calling Save.
-    let resolutions: Vec<(crate::services::sportify::models::SportifyTrack, i64)> = db
-        .with_conn(|conn| {
-            let mut out = Vec::new();
-            for t in &playlist.tracks {
-                let Some(spotify_track_id) = t.id.as_deref() else {
-                    continue;
-                };
-                if let Some(hit) =
-                    sp_cache::get_tidal_resolution(conn, &cache_cfg, spotify_track_id)?
-                {
-                    out.push((t.clone(), hit.tidal_track_id));
-                }
-            }
-            Ok::<_, anyhow::Error>(out)
-        })
-        .map_err(internal)?;
-
-    let total_tracks = playlist.tracks.len();
-    let resolved_count = resolutions.len();
-    let unresolved_count = total_tracks.saturating_sub(resolved_count);
-
-    if resolutions.is_empty() {
-        return Err((
-            StatusCode::UNPROCESSABLE_ENTITY,
-            Json(json!({
-                "error": "no_resolved_tracks",
-                "totalTracks": total_tracks,
-                "resolvedCount": 0,
-                "unresolvedCount": unresolved_count,
-            })),
-        ));
-    }
-
-    // Import each resolved TIDAL track into the local library so it has a
-    // local_id for the playlist join. Failures are skipped, not fatal.
-    let mut local_ids: Vec<i64> = Vec::with_capacity(resolutions.len());
-    let mut import_failures: usize = 0;
-    for (sp_track, tidal_id) in &resolutions {
-        let metadata = tidal_import::ImportTrackMetadata {
-            tidal_id: *tidal_id,
-            title: sp_track.name.clone().unwrap_or_default(),
-            artist_name: sp_track
-                .primary_artist()
-                .map(str::to_string)
-                .unwrap_or_default(),
-            artist_tidal_id: None,
-            artist_picture: None,
-            album_title: sp_track.album.as_ref().and_then(|a| a.name.clone()),
-            album_tidal_id: None,
-            album_artwork_url: sp_track.best_thumbnail(),
-            duration_ms: sp_track.duration_ms,
-        };
-        match tidal_import::import_track_from_metadata(&db, metadata).await {
-            Ok(imported) => local_ids.push(imported.local_id),
-            Err(e) => {
-                tracing::warn!(
-                    "save_spotify_playlist: import failed for tidal_id {}: {}",
-                    tidal_id,
-                    e
-                );
-                import_failures += 1;
-            }
-        }
-    }
-
-    if local_ids.is_empty() {
-        return Err((
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(json!({
-                "error": "all_imports_failed",
-                "totalTracks": total_tracks,
-                "resolvedCount": resolved_count,
-                "importFailures": import_failures,
-            })),
-        ));
-    }
-
-    // Create the playlist row and bulk-add tracks in a single transaction.
-    let result = db
-        .with_conn(|conn| {
-            conn.execute(
-                "INSERT INTO playlists (name, description, is_smart, is_synced, track_count)
-                 VALUES (?1, ?2, 0, 0, 0)",
-                params![playlist_name, playlist.description],
-            )?;
-            let playlist_id = conn.last_insert_rowid();
-            let added = queries::add_tracks_to_playlist(conn, playlist_id, &local_ids)?;
-            let row = queries::get_playlist(conn, playlist_id)?
-                .ok_or_else(|| anyhow::anyhow!("playlist not found after insert"))?;
-            Ok::<_, anyhow::Error>((row, added))
-        })
-        .map_err(internal)?;
-
-    Ok(Json(json!({
-        "playlist": result.0,
-        "added": result.1,
-        "totalTracks": total_tracks,
-        "resolvedCount": resolved_count,
-        "unresolvedCount": unresolved_count,
-        "importFailures": import_failures,
-    })))
 }
 
 fn internal<E: std::fmt::Display>(e: E) -> (StatusCode, Json<Value>) {
@@ -15609,712 +14757,6 @@ async fn reanalyze_stale_tracks(
             "Analysis actor not configured. Stale rows cleared but no scan queued."
         }
     })))
-}
-
-// ─── Trending / Charts (Phase 5) ────────────────────────────────────────────
-
-#[derive(Debug, Deserialize)]
-pub struct ChartParams {
-    /// "lastfm" (default) or "tidal".
-    source: Option<String>,
-    /// Max entries to return (clamped 1..=100).
-    limit: Option<u32>,
-    /// Optional country (Last.fm only). Accepts either an ISO 3166-1 alpha-2
-    /// code (e.g. "AU") which is mapped via `CURATED_COUNTRIES`, or the full
-    /// English name (e.g. "United States") for legacy/free-form callers.
-    country: Option<String>,
-    /// Optional curated genre key (Last.fm only), e.g. "hip-hop".
-    /// Mutually exclusive with `country`.
-    tag: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct ChartTidalPlayable {
-    tidal_id: i64,
-    title: String,
-    artist_name: Option<String>,
-    artist_tidal_id: Option<i64>,
-    album_title: Option<String>,
-    artwork_url: Option<String>,
-    duration_ms: Option<i64>,
-    track_id: Option<i64>,
-    is_in_library: bool,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct ChartEntryDto {
-    /// Local library Track when the chart entry was resolved to a known track.
-    /// Frontend renders these via `<TrackRow>` and gets full menu support.
-    local_track: Option<crate::db::models::Track>,
-    /// Otherwise a TidalPlayable-shaped DTO that the frontend renders via
-    /// `<TidalTrackRow>`. May be `None` only when both resolutions failed (rare).
-    tidal_playable: Option<ChartTidalPlayable>,
-    /// Optional preview image (mostly for Last.fm entries with no Tidal match).
-    image_url: Option<String>,
-    /// Source-tagged for the frontend ("lastfm" | "tidal").
-    source: String,
-    /// Top genre name for the resolved local track (None for Tidal-only entries
-    /// where we have no genre data without an extra API call).
-    genre: Option<String>,
-}
-
-/// Fill in missing artwork for chart entries by searching Tidal for the
-/// (artist, title) pair. Updates `image_url` (top-level fallback) and the
-/// nested `tidal_playable.artwork_url` so the frontend's preference chain
-/// always lands on something usable.
-async fn enrich_chart_artwork(state: &SharedState, entries: &mut Vec<ChartEntryDto>) {
-    use futures::stream::{FuturesUnordered, StreamExt};
-
-    /// Last.fm's blank-star fallback. We never want to surface this — both the
-    /// usable-art check below and the replace-on-enrich path treat it as empty.
-    const LASTFM_PLACEHOLDER: &str = "2a96cbd8b46e442fc41c2b86b821562f";
-
-    fn is_unusable(url: Option<&str>) -> bool {
-        let Some(url) = url else { return true };
-        let trimmed = url.trim();
-        trimmed.is_empty() || trimmed.contains(LASTFM_PLACEHOLDER)
-    }
-
-    fn has_usable_art(e: &ChartEntryDto) -> bool {
-        let local_ok = !is_unusable(
-            e.local_track
-                .as_ref()
-                .and_then(|t| t.artwork_url.as_deref()),
-        );
-        let tp_ok = !is_unusable(
-            e.tidal_playable
-                .as_ref()
-                .and_then(|tp| tp.artwork_url.as_deref()),
-        );
-        let img_ok = !is_unusable(e.image_url.as_deref());
-        local_ok || tp_ok || img_ok
-    }
-
-    let needs: Vec<(usize, String, String)> = entries
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, e)| {
-            if has_usable_art(e) {
-                return None;
-            }
-            let title = e
-                .local_track
-                .as_ref()
-                .map(|t| t.title.clone())
-                .or_else(|| e.tidal_playable.as_ref().map(|tp| tp.title.clone()))?;
-            let artist = e
-                .local_track
-                .as_ref()
-                .and_then(|t| t.artist_name.clone())
-                .or_else(|| {
-                    e.tidal_playable
-                        .as_ref()
-                        .and_then(|tp| tp.artist_name.clone())
-                })?;
-            Some((idx, artist, title))
-        })
-        .collect();
-
-    if needs.is_empty() {
-        return;
-    }
-
-    let (tokens, tidal_http_client) = {
-        let s = state.read().await;
-        (s.tidal_tokens.clone(), s.tidal_http_client.clone())
-    };
-    let tokens = match tokens {
-        Some(t) => Some(t),
-        None => load_persisted_tidal_tokens(state).await.ok().flatten(),
-    };
-    let Some(tokens) = tokens else {
-        return;
-    };
-
-    let client = TidalClient::with_http(
-        tidal_http_client,
-        tokens.access_token.clone(),
-        tokens.country_code.clone(),
-    );
-
-    let mut tasks = FuturesUnordered::new();
-    for (idx, artist, title) in needs {
-        let client = client.clone();
-        tasks.push(async move {
-            let q = format!("{artist} {title}");
-            let result = client.search(&q, 1).await.ok();
-            let url = result
-                .and_then(|r| r.into_iter().next())
-                .and_then(|t| t.artwork_url);
-            (idx, url)
-        });
-    }
-
-    while let Some((idx, url)) = tasks.next().await {
-        let Some(url) = url else { continue };
-        if let Some(entry) = entries.get_mut(idx) {
-            if is_unusable(entry.image_url.as_deref()) {
-                entry.image_url = Some(url.clone());
-            }
-            if let Some(tp) = entry.tidal_playable.as_mut()
-                && is_unusable(tp.artwork_url.as_deref())
-            {
-                tp.artwork_url = Some(url);
-            }
-        }
-    }
-}
-
-/// Look up the most-confident genre name for each track id in a single query.
-/// Returns a map keyed by track_id; tracks with no genre rows are absent.
-fn fetch_top_genres_for_tracks(
-    db: &crate::db::Database,
-    track_ids: &[i64],
-) -> HashMap<i64, String> {
-    if track_ids.is_empty() {
-        return HashMap::new();
-    }
-    db.with_conn(|conn| {
-        let placeholders = std::iter::repeat_n("?", track_ids.len())
-            .collect::<Vec<_>>()
-            .join(",");
-        // Pick the highest-confidence genre per track. Ties broken by
-        // alphabetical order so the result is stable.
-        let sql = format!(
-            "SELECT track_id, name FROM (
-                SELECT tg.track_id, g.name,
-                       ROW_NUMBER() OVER (
-                           PARTITION BY tg.track_id
-                           ORDER BY tg.confidence DESC, g.name ASC
-                       ) AS rn
-                FROM track_genres tg
-                JOIN genres g ON g.id = tg.genre_id
-                WHERE tg.track_id IN ({placeholders})
-             ) WHERE rn = 1"
-        );
-        let mut stmt = conn.prepare(&sql)?;
-        let params: Vec<&dyn rusqlite::ToSql> = track_ids
-            .iter()
-            .map(|id| id as &dyn rusqlite::ToSql)
-            .collect();
-        let rows = stmt.query_map(params.as_slice(), |row| {
-            Ok((row.get::<_, i64>(0)?, row.get::<_, String>(1)?))
-        })?;
-        let mut map = HashMap::new();
-        for row in rows {
-            let (id, name) = row?;
-            map.insert(id, name);
-        }
-        Ok(map)
-    })
-    .unwrap_or_default()
-}
-
-/// Cached chart payload with insertion timestamp.
-struct ChartCacheEntry {
-    inserted_at: Instant,
-    payload: serde_json::Value,
-}
-
-fn chart_cache() -> &'static StdMutex<HashMap<String, ChartCacheEntry>> {
-    static CACHE: OnceLock<StdMutex<HashMap<String, ChartCacheEntry>>> = OnceLock::new();
-    CACHE.get_or_init(|| StdMutex::new(HashMap::new()))
-}
-
-/// Charts don't move minute-to-minute and Last.fm rate-limits aggressively.
-/// 2-hour TTL matches the frontend's in-memory cache so the trending shelf
-/// stays static across page navigations within the window.
-const CHART_CACHE_TTL: Duration = Duration::from_secs(2 * 60 * 60);
-
-fn chart_cache_get(key: &str) -> Option<serde_json::Value> {
-    let cache = chart_cache().lock().ok()?;
-    let entry = cache.get(key)?;
-    if entry.inserted_at.elapsed() > CHART_CACHE_TTL {
-        return None;
-    }
-    Some(entry.payload.clone())
-}
-
-fn chart_cache_put(key: String, payload: serde_json::Value) {
-    if let Ok(mut cache) = chart_cache().lock() {
-        // Bound cache size to avoid unbounded growth from arbitrary
-        // source/country/limit combos. 32 entries is plenty.
-        if cache.len() >= 32 {
-            cache.clear();
-        }
-        cache.insert(
-            key,
-            ChartCacheEntry {
-                inserted_at: Instant::now(),
-                payload,
-            },
-        );
-    }
-}
-
-async fn get_charts(
-    State(state): State<SharedState>,
-    Query(params): Query<ChartParams>,
-) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    use crate::services::charts::curated;
-
-    let source = params
-        .source
-        .as_deref()
-        .map(|s| s.trim().to_ascii_lowercase())
-        .filter(|s| !s.is_empty())
-        .unwrap_or_else(|| "lastfm".to_string());
-    let limit = params.limit.unwrap_or(50).clamp(1, 100);
-
-    let country_input = params
-        .country
-        .as_deref()
-        .map(str::trim)
-        .filter(|s| !s.is_empty());
-    let tag_input = params
-        .tag
-        .as_deref()
-        .map(str::trim)
-        .filter(|s| !s.is_empty());
-
-    if country_input.is_some() && tag_input.is_some() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(json!({ "error": "tag_country_exclusive" })),
-        ));
-    }
-
-    // Resolve country: ISO code or full name → curated entry. Two-char inputs
-    // that aren't curated codes are rejected; longer strings that don't match
-    // a curated `lastfm_name` pass through as free-form (legacy callers).
-    // The cache token is always the ISO code when curated, so `?country=AU`
-    // and `?country=Australia` collapse to one cache entry.
-    let (country_resolved, country_cache_token): (Option<String>, Option<String>) =
-        match country_input {
-            None => (None, None),
-            Some(s) if s.len() == 2 => match curated::find_country_by_code(s) {
-                Some(entry) => (
-                    Some(entry.lastfm_name.to_string()),
-                    Some(entry.code.to_string()),
-                ),
-                None => {
-                    return Err((
-                        StatusCode::BAD_REQUEST,
-                        Json(json!({ "error": "unknown_country" })),
-                    ));
-                }
-            },
-            Some(s) => match curated::find_country_by_code_or_name(s) {
-                Some(entry) => (
-                    Some(entry.lastfm_name.to_string()),
-                    Some(entry.code.to_string()),
-                ),
-                None => (Some(s.to_string()), Some(s.to_ascii_uppercase())),
-            },
-        };
-
-    let tag_resolved: Option<&'static curated::GenreEntry> = match tag_input {
-        Some(key) => match curated::find_genre(key) {
-            Some(entry) => Some(entry),
-            None => {
-                return Err((
-                    StatusCode::BAD_REQUEST,
-                    Json(json!({ "error": "unknown_genre" })),
-                ));
-            }
-        },
-        None => None,
-    };
-
-    let cache_key = format!(
-        "{}|{}|{}|{}",
-        source,
-        limit,
-        country_cache_token.as_deref().unwrap_or(""),
-        tag_resolved.map(|g| g.key).unwrap_or("")
-    );
-    if let Some(cached) = chart_cache_get(&cache_key) {
-        return Ok(Json(cached));
-    }
-
-    let mut entries: Vec<ChartEntryDto> = match source.as_str() {
-        "tidal" => fetch_tidal_chart(&state, limit as i32).await.map_err(|e| {
-            (
-                StatusCode::BAD_GATEWAY,
-                Json(json!({ "error": format!("tidal chart: {e}") })),
-            )
-        })?,
-        _ => fetch_lastfm_chart(&state, limit, country_resolved.as_deref(), tag_resolved)
-            .await
-            .map_err(|e| {
-                (
-                    StatusCode::BAD_GATEWAY,
-                    Json(json!({ "error": format!("lastfm chart: {e}") })),
-                )
-            })?,
-    };
-
-    // Backfill missing artwork via Tidal search. Last.fm's chart.getTopTracks
-    // mostly returns a generic placeholder image, and many older library albums
-    // have NULL artwork_url, so this is the difference between blank tiles and
-    // real covers on the trending shelf.
-    enrich_chart_artwork(&state, &mut entries).await;
-
-    let payload = json!({
-        "source": source,
-        "limit": limit,
-        "country": country_cache_token,
-        "tag": tag_resolved.map(|g| g.key),
-        "tracks": entries,
-    });
-    chart_cache_put(cache_key, payload.clone());
-    Ok(Json(payload))
-}
-
-async fn list_lastfm_genres() -> Json<Value> {
-    use crate::services::charts::curated::{CURATED_GENRES, DEFAULT_GENRE_KEY};
-    let genres: Vec<_> = CURATED_GENRES
-        .iter()
-        .map(|g| json!({ "key": g.key, "label": g.label }))
-        .collect();
-    Json(json!({ "genres": genres, "default_genre": DEFAULT_GENRE_KEY }))
-}
-
-async fn list_lastfm_countries() -> Json<Value> {
-    use crate::services::charts::curated::{CURATED_COUNTRIES, DEFAULT_COUNTRY_CODE};
-    let countries: Vec<_> = CURATED_COUNTRIES
-        .iter()
-        .map(|c| json!({ "code": c.code, "label": c.label }))
-        .collect();
-    Json(json!({ "countries": countries, "default_country": DEFAULT_COUNTRY_CODE }))
-}
-
-async fn fetch_lastfm_chart(
-    state: &SharedState,
-    limit: u32,
-    country: Option<&str>,
-    genre: Option<&'static crate::services::charts::curated::GenreEntry>,
-) -> anyhow::Result<Vec<ChartEntryDto>> {
-    let (http, db) = {
-        let s = state.read().await;
-        (s.http_client.clone(), s.db.clone())
-    };
-    let client = LastFmClient::load(http, &db)
-        .ok_or_else(|| anyhow::anyhow!("Last.fm API key not configured"))?;
-
-    let tracks = if let Some(genre) = genre {
-        // Genre tag — fan out to every Last.fm tag the curated entry maps to,
-        // merge, dedupe by normalised (artist, title), sum playcounts on dupes,
-        // sort desc by playcount, truncate. Single-tag entries skip the merge.
-        if genre.lastfm_tags.len() == 1 {
-            client
-                .get_top_tracks_by_tag(genre.lastfm_tags[0], limit)
-                .await?
-        } else {
-            use futures::future::join_all;
-            // Overfetch per leg so dedup across overlapping tags doesn't shrink
-            // the merged list below the requested `limit`. Capped at Last.fm's
-            // per-call ceiling.
-            let fan_limit = limit.saturating_mul(2).min(100);
-            let calls = genre.lastfm_tags.iter().map(|tag| {
-                let c = client.clone();
-                let t = (*tag).to_string();
-                async move { c.get_top_tracks_by_tag(&t, fan_limit).await }
-            });
-            let results = join_all(calls).await;
-            let mut merged: Vec<crate::metadata::lastfm::LastFmChartTrack> = Vec::new();
-            let mut by_key: HashMap<String, usize> = HashMap::new();
-            for res in results {
-                let list = match res {
-                    Ok(l) => l,
-                    Err(e) => {
-                        tracing::warn!("tag fan-out leg failed: {}", e);
-                        continue;
-                    }
-                };
-                for t in list {
-                    let key = crate::services::radio::normalize_for_dedup(&t.artist, &t.title);
-                    if key.is_empty() {
-                        continue;
-                    }
-                    if let Some(&idx) = by_key.get(&key) {
-                        // Merge: sum playcounts/listeners; prefer the first non-empty
-                        // image and mbid (already populated on the existing entry).
-                        let existing: &mut crate::metadata::lastfm::LastFmChartTrack =
-                            &mut merged[idx];
-                        existing.playcount = match (existing.playcount, t.playcount) {
-                            (Some(a), Some(b)) => Some(a.saturating_add(b)),
-                            (a, b) => a.or(b),
-                        };
-                        existing.listeners = match (existing.listeners, t.listeners) {
-                            (Some(a), Some(b)) => Some(a.saturating_add(b)),
-                            (a, b) => a.or(b),
-                        };
-                        if existing.image_url.as_deref().unwrap_or("").is_empty() {
-                            existing.image_url = t.image_url;
-                        }
-                        if existing.mbid.is_none() {
-                            existing.mbid = t.mbid;
-                        }
-                    } else {
-                        by_key.insert(key, merged.len());
-                        merged.push(t);
-                    }
-                }
-            }
-            merged.sort_by(|a, b| b.playcount.unwrap_or(0).cmp(&a.playcount.unwrap_or(0)));
-            merged.truncate(limit as usize);
-            merged
-        }
-    } else {
-        client.get_top_chart(limit, country).await?
-    };
-
-    // Resolve each (artist, title) to a local library track when present.
-    // We do this in a single DB call by collecting all (artist, title) pairs
-    // and matching case-insensitively.
-    let pairs: Vec<(String, String)> = tracks
-        .iter()
-        .map(|t| (t.artist.clone(), t.title.clone()))
-        .collect();
-    let local_map = resolve_chart_pairs_to_local(&db, &pairs).unwrap_or_default();
-    let local_ids: Vec<i64> = local_map.values().map(|t| t.id).collect();
-    let genre_map = fetch_top_genres_for_tracks(&db, &local_ids);
-
-    let mut out = Vec::with_capacity(tracks.len());
-    for t in tracks {
-        let key = format!(
-            "{}\u{0001}{}",
-            t.artist.to_ascii_lowercase(),
-            t.title.to_ascii_lowercase()
-        );
-        let local_track = local_map.get(&key).cloned();
-        let genre = local_track
-            .as_ref()
-            .and_then(|lt| genre_map.get(&lt.id).cloned());
-        let tidal_playable = if local_track.is_none() {
-            // No local match; expose a TidalPlayable-shaped placeholder. The
-            // frontend will resolve to a real Tidal id via search if the user
-            // clicks play (existing ephemeral-play flow does this).
-            Some(ChartTidalPlayable {
-                tidal_id: 0,
-                title: t.title.clone(),
-                artist_name: Some(t.artist.clone()),
-                artist_tidal_id: None,
-                album_title: None,
-                artwork_url: t.image_url.clone(),
-                duration_ms: None,
-                track_id: None,
-                is_in_library: false,
-            })
-        } else {
-            None
-        };
-        out.push(ChartEntryDto {
-            local_track,
-            tidal_playable,
-            image_url: t.image_url,
-            source: "lastfm".to_string(),
-            genre,
-        });
-    }
-    Ok(out)
-}
-
-async fn fetch_tidal_chart(state: &SharedState, limit: i32) -> anyhow::Result<Vec<ChartEntryDto>> {
-    let (tokens_opt, http, db, tidal_http_client) = {
-        let s = state.read().await;
-        (
-            s.tidal_tokens.clone(),
-            s.http_client.clone(),
-            s.db.clone(),
-            s.tidal_http_client.clone(),
-        )
-    };
-    let persisted = load_persisted_tidal_tokens(state).await?;
-    let tokens = tokens_opt.or(persisted);
-    let Some(tokens) = tokens else {
-        // Tidal not connected; degrade to empty list rather than failing.
-        tracing::warn!("Tidal chart requested but Tidal not connected");
-        return Ok(Vec::new());
-    };
-    let client = TidalClient::with_http(
-        tidal_http_client,
-        tokens.access_token.clone(),
-        tokens.country_code.clone(),
-    );
-    let tracks = match client.get_editorial_top_tracks(limit).await {
-        Ok(t) => t,
-        Err(e) => {
-            tracing::warn!("Tidal editorial chart failed: {}", e);
-            Vec::new()
-        }
-    };
-    if tracks.is_empty() {
-        return Ok(Vec::new());
-    }
-    let _ = http; // currently unused; reserved for future fallback paths
-
-    // Resolve to local tracks via tidal_id batch lookup.
-    let tidal_ids: Vec<i64> = tracks.iter().map(|t| t.id).collect();
-    let known: HashMap<i64, i64> = db
-        .with_conn(|conn| queries::get_tidal_track_local_ids(conn, &tidal_ids))
-        .unwrap_or_default();
-
-    // Pull full Track rows for any local matches.
-    let local_ids: Vec<i64> = known.values().copied().collect();
-    let local_tracks: HashMap<i64, crate::db::models::Track> = db
-        .with_conn(|conn| {
-            let mut map: HashMap<i64, crate::db::models::Track> = HashMap::new();
-            if local_ids.is_empty() {
-                return Ok(map);
-            }
-            let mut stmt = conn.prepare(
-                "SELECT t.id, t.title, t.artist_id, a.name as artist_name, t.album_id, al.title as album_title,
-                        t.disc_number, t.track_number, t.duration_ms, t.isrc, t.tidal_id, t.ytmusic_id,
-                        t.soundcloud_id, t.best_quality, t.best_source, t.fidelity_score, t.is_favorite,
-                        t.play_count, t.last_played_at, t.date_added, t.source, t.artwork_url
-                 FROM tracks t
-                 LEFT JOIN artists a ON t.artist_id = a.id
-                 LEFT JOIN albums al ON t.album_id = al.id
-                 WHERE t.id = ?1
-                 LIMIT 1",
-            )?;
-            for id in &local_ids {
-                let mut rows = stmt.query(rusqlite::params![id])?;
-                if let Some(row) = rows.next()? {
-                    let track = crate::db::models::Track {
-                        id: row.get(0)?,
-                        title: row.get(1)?,
-                        artist_id: row.get(2)?,
-                        artist_name: row.get(3)?,
-                        album_id: row.get(4)?,
-                        album_title: row.get(5)?,
-                        disc_number: row.get(6)?,
-                        track_number: row.get(7)?,
-                        duration_ms: row.get(8)?,
-                        isrc: row.get(9)?,
-                        tidal_id: row.get(10)?,
-                        ytmusic_id: row.get(11)?,
-                        soundcloud_id: row.get(12)?,
-                        best_quality: row.get(13)?,
-                        best_source: row.get(14)?,
-                        fidelity_score: row.get(15)?,
-                        is_favorite: row.get::<_, i64>(16)? != 0,
-                        play_count: row.get(17)?,
-                        last_played_at: row.get(18)?,
-                        date_added: row.get(19)?,
-                        source: row.get(20)?,
-                        artwork_url: row.get(21)?,
-                    };
-                    map.insert(*id, track);
-                }
-            }
-            Ok(map)
-        })
-        .unwrap_or_default();
-
-    let resolved_local_ids: Vec<i64> = local_tracks.values().map(|t| t.id).collect();
-    let genre_map = fetch_top_genres_for_tracks(&db, &resolved_local_ids);
-
-    let mut out = Vec::with_capacity(tracks.len());
-    for t in tracks {
-        let local_track = known
-            .get(&t.id)
-            .and_then(|lid| local_tracks.get(lid))
-            .cloned();
-        let genre = local_track
-            .as_ref()
-            .and_then(|lt| genre_map.get(&lt.id).cloned());
-        let tidal_playable = if local_track.is_none() {
-            Some(ChartTidalPlayable {
-                tidal_id: t.id,
-                title: t.title.clone(),
-                artist_name: t.artist_name.clone(),
-                artist_tidal_id: t.artist_id,
-                album_title: t.album_title.clone(),
-                artwork_url: t.artwork_url.clone(),
-                duration_ms: Some(t.duration * 1000),
-                track_id: None,
-                is_in_library: false,
-            })
-        } else {
-            None
-        };
-        out.push(ChartEntryDto {
-            image_url: t.artwork_url.clone(),
-            local_track,
-            tidal_playable,
-            source: "tidal".to_string(),
-            genre,
-        });
-    }
-    Ok(out)
-}
-
-/// Resolve (artist, title) pairs to local Track rows, case-insensitively.
-/// Uses a single SQL query with a fold-table; falls back to empty map on error.
-fn resolve_chart_pairs_to_local(
-    db: &crate::db::Database,
-    pairs: &[(String, String)],
-) -> anyhow::Result<HashMap<String, crate::db::models::Track>> {
-    if pairs.is_empty() {
-        return Ok(HashMap::new());
-    }
-    let mut out = HashMap::new();
-    db.with_conn(|conn| {
-        // Match by lower(artist_name) + lower(title). Library is small enough
-        // that one round-trip per pair is acceptable; a single OR-chained
-        // query also works but is harder to map back to pairs.
-        let mut stmt = conn.prepare(
-            "SELECT t.id, t.title, t.artist_id, a.name as artist_name, t.album_id, al.title as album_title,
-                    t.disc_number, t.track_number, t.duration_ms, t.isrc, t.tidal_id, t.ytmusic_id,
-                    t.soundcloud_id, t.best_quality, t.best_source, t.fidelity_score, t.is_favorite,
-                    t.play_count, t.last_played_at, t.date_added, t.source, t.artwork_url
-             FROM tracks t
-             LEFT JOIN artists a ON t.artist_id = a.id
-             LEFT JOIN albums al ON t.album_id = al.id
-             WHERE LOWER(a.name) = LOWER(?1) AND LOWER(t.title) = LOWER(?2)
-             LIMIT 1",
-        )?;
-        for (artist, title) in pairs {
-            let mut rows = stmt.query(rusqlite::params![artist, title])?;
-            if let Some(row) = rows.next()? {
-                let track = crate::db::models::Track {
-                    id: row.get(0)?,
-                    title: row.get(1)?,
-                    artist_id: row.get(2)?,
-                    artist_name: row.get(3)?,
-                    album_id: row.get(4)?,
-                    album_title: row.get(5)?,
-                    disc_number: row.get(6)?,
-                    track_number: row.get(7)?,
-                    duration_ms: row.get(8)?,
-                    isrc: row.get(9)?,
-                    tidal_id: row.get(10)?,
-                    ytmusic_id: row.get(11)?,
-                    soundcloud_id: row.get(12)?,
-                    best_quality: row.get(13)?,
-                    best_source: row.get(14)?,
-                    fidelity_score: row.get(15)?,
-                    is_favorite: row.get::<_, i64>(16)? != 0,
-                    play_count: row.get(17)?,
-                    last_played_at: row.get(18)?,
-                    date_added: row.get(19)?,
-                    source: row.get(20)?,
-                    artwork_url: row.get(21)?,
-                };
-                let key = format!(
-                    "{}\u{0001}{}",
-                    artist.to_ascii_lowercase(),
-                    title.to_ascii_lowercase()
-                );
-                out.insert(key, track);
-            }
-        }
-        Ok(())
-    })?;
-    Ok(out)
 }
 
 #[cfg(test)]
